Move Publicize connection URL matching to its own file
